requirements.txt
alembic==1.8.0
amqp==5.1.1
async-timeout==4.0.2
Babel==2.10.3
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.2.1
dominate==2.6.0
elastic-transport==8.1.2
elasticsearch==8.3.1
email-validator==1.2.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Babel==2.0.0
Flask-BasicAuth==0.2.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
guess-language-spirit==0.5.3
idna==3.3
importlib-metadata==4.11.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
langdetect==1.0.9
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
